# apps/index/src/routes/challenge.ts
import { nsName } from '../lib/names';
import { getSite, recordAudit } from '@identizen/db';
import { AcrSchema, REASON_MAX_LENGTH, type Acr } from '@identizen/protocol';
import { Hono } from 'hono';
import { z } from 'zod';
import type { AppEnv } from '../app';
import { isReservedError, type OidcParams } from '../do/challenge-session';
import { badRequest, conflict, forbidden, notFound } from '../lib/errors';
import { registeredRedirect, validateAuthorizeRequest } from '../oidc/authorize-request';
import { browserMeta } from '../lib/util';
import { deviceAuth } from '../middleware/idz-signature';
import { ipRateLimit } from '../middleware/rate-limit';
import { processAssertion } from '../services/assert';
import { completeApproval, onSessionDenied } from '../services/sessions';
import { startChallenge } from '../services/challenge';

const StartSchema = z
  .object({
    client_id: z.string().min(1),
    acr: AcrSchema.default('idz:login'),
    reason: z.string().min(1).max(REASON_MAX_LENGTH).nullable().optional(),
    login_hint: z.string().min(1).optional(),
    browser_pubkey: z
      .string()
      .regex(/^[A-Za-z0-9_-]+$/)
      .optional(),
    redirect_uri: z.string().url().optional(),
    state: z.string().max(512).optional(),
    nonce: z.string().max(512).optional(),
    code_challenge: z.string().min(43).max(128).optional(),
    code_challenge_method: z.literal('S256').optional(),
    scope: z.string().max(256).optional(),
    prompt: z.enum(['enroll', 'login', 'none', 'consent']).optional(),
  })
  .strict();

export function challengeRoutes(): Hono<AppEnv> {
  const r = new Hono<AppEnv>();

  /**
   * Start a login (the JSON twin of `/authorize`; the SDK and `/authorize` both use it).
   * Returns the ids the browser needs; the phone fetches the signed challenge itself.
   */
  r.post('/challenge', ipRateLimit(), async (c) => {
    const services = c.get('services');
    const body = StartSchema.parse(await c.req.json());
    // With a redirect_uri this is an authorization request and is held to exactly the rules
    // `/authorize` applies: registered redirect, PKCE, known scopes, prompt semantics. Errors
    // come back as JSON rather than a redirect, because the caller is the site's own code.
    let acr: Acr = body.acr;
    let loginHint: string | null = body.login_hint ?? null;
    let oidc: OidcParams | null = null;
    if (body.redirect_uri !== undefined) {
      const site = await getSite(services.db, body.client_id);
      if (!site) throw notFound('unknown_client', `no site with client_id ${body.client_id}`);
      const redirectUri = registeredRedirect(site, body.redirect_uri);
      if (!redirectUri)
        throw badRequest('invalid_request', 'redirect_uri is not registered for this client');
      const v = validateAuthorizeRequest(
        c.env,
        site,
        {
          scope: body.scope ?? 'openid',
          state: body.state,
          nonce: body.nonce,
          code_challenge: body.code_challenge,
          code_challenge_method: body.code_challenge_method,
          acr_values: body.acr,
          prompt: body.prompt,
          login_hint: body.login_hint,
        },
        { redirectUri, responseTypeRequired: false },
      );
      if (!v.ok) throw badRequest(v.error, v.description);
      acr = v.acr;
      loginHint = v.loginHint;
      oidc = v.oidc;
    }
    const result = await startChallenge(
      services,
      {
        clientId: body.client_id,
        acr,
        reason: body.reason ?? null,
        loginHint,
        browserPubkey: body.browser_pubkey ?? null,
        browser: browserMeta(c),
        oidc,
      },
      c.env,
    );
    const id = result.signed.payload.id;
    const wsUrl = new URL(`/challenge/${id}/ws`, services.indexUrl);
    wsUrl.protocol = wsUrl.protocol === 'https:' ? 'wss:' : 'ws:';
    return c.json(
      {
        challenge_id: id,
        code: result.signed.payload.code,
        exp: result.signed.payload.exp,
        acr: result.signed.payload.acr,
        rp_name: result.site.name,
        deep_link: `${services.appUrl}/l/${id}?index=${encodeURIComponent(services.indexUrl)}`,
        ws_url: wsUrl.toString(),
        pushed: result.pushedTo !== null,
      },
      201,
    );
  });

  /** Phone fetches the signed challenge (after push, deep link, or QR). */
  r.get('/challenge/:id', async (c) => {
    const stub = c.get('services').stores.challenge(nsName(c.env, c.req.param('id')));
    const [signed, state] = await Promise.all([stub.getSigned(), stub.getState()]);
    if (!signed || !state) throw notFound('unknown_challenge', 'no such challenge');
    return c.json({ ...signed, status: state.status });
  });

  /** Browser polls (fallback for environments without WebSocket). */
  r.get('/challenge/:id/state', async (c) => {
    const stub = c.get('services').stores.challenge(nsName(c.env, c.req.param('id')));
    const state = await stub.getState();
    if (!state) throw notFound('unknown_challenge', 'no such challenge');
    return c.json({
      challenge_id: state.challengeId,
      status: state.status,
      pairing: state.pairing,
      redirect: state.status === 'approved' ? state.redirect : null,
    });
  });

  /** Hosted page: attach the browser P-256 key so approval issues a pairing. */
  r.post('/challenge/:id/browser-key', async (c) => {
    const body = z
      .object({ browser_pubkey: z.string().regex(/^[A-Za-z0-9_-]{80,100}$/) })
      .strict()
      .parse(await c.req.json());
    const stub = c.get('services').stores.challenge(nsName(c.env, c.req.param('id')));
    const state = await stub.getState();
    if (!state) throw notFound('unknown_challenge', 'no such challenge');
    const ok = await stub.setBrowserPubkey(body.browser_pubkey, browserMeta(c));
    return c.json({ ok });
  });

  /** Browser WebSocket to the session store (the page polls `/state` if the store has none). */
  r.get('/challenge/:id/ws', (c) => {
    const stub = c.get('services').stores.challenge(nsName(c.env, c.req.param('id')));
    return stub.websocket(c.req.raw);
  });

  /** Phone submits the signed assertion. Signed request (device auth) + double-signed body. */
  r.post('/challenge/:id/assert', deviceAuth(), async (c) => {
    const services = c.get('services');
    const id = c.req.param('id');
    const stub = services.stores.challenge(nsName(c.env, id));
    const body: unknown = JSON.parse(c.get('rawBody') || '{}');
    const caller = c.get('device');
    const claimed = (body as { payload?: { device_id?: string } }).payload?.device_id;
    if (claimed !== caller.id)
      throw forbidden('wrong_device', 'assertion device_id does not match the signing device');
    const outcome = await processAssertion(services, stub, id, body);
    const completed = await completeApproval(services, c.env, stub, outcome);
    return c.json({
      status: 'approved',
      challenge_id: id,
      sub: outcome.assertion.sub,
      acr: outcome.assertion.acr,
      pairing: outcome.pairing,
      redirect: completed.redirect,
    });
  });

  /** Phone declines. */
  r.post('/challenge/:id/deny', deviceAuth(), async (c) => {
    const { db, stores } = c.get('services');
    const id = c.req.param('id');
    const stub = stores.challenge(nsName(c.env, id));
    const state = await stub.getState();
    if (!state) throw notFound('unknown_challenge', 'no such challenge');
    if (state.status !== 'pending') return c.json({ status: state.status, challenge_id: id });
    const device = c.get('device');
    // Only the person the challenge was issued for may decline it (any of their devices); an
    // untargeted challenge that discovery routed to one device is that device's to decline.
    if (state.expectedIdz !== null) {
      if (device.idz !== state.expectedIdz)
        throw forbidden('wrong_identity', 'this challenge was issued to a different identity');
    } else if (state.targetDeviceId !== null && state.targetDeviceId !== device.id) {
      throw forbidden('wrong_device', 'this challenge was routed to a different device');
    }
    try {
      await stub.deny();
    } catch (err) {
      // An approval already holds the session (F03): the assertion verified and its writes are
      // in flight, so this decline arrives too late to change the outcome.
      if (isReservedError(err))
        throw conflict('approval_in_progress', 'this challenge is being approved');
      throw err;
    }
    await recordAudit(db, {
      kind: 'login.denied',
      idz: device.idz,
      deviceId: device.id,
      clientId: state.clientId,
      detail: { challenge_id: id, reason: 'user_denied' },
    });
    await onSessionDenied(c.get('services'), state);
    return c.json({ status: 'denied', challenge_id: id });
  });

  return r;
}