# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Identizen index API
  version: 0.1.0
  summary: The Identizen index — device registry, login relay, and OpenID Provider.
  description: |
    Machine-readable description of the HTTP API served by an Identizen index
    (`apps/index`). It is written by hand from the Hono routes in
    `apps/index/src/routes/*.ts`, the Zod schemas in `packages/protocol/src/schemas.ts`,
    and the error helpers in `apps/index/src/lib/errors.ts`, and it is validated by
    `apps/index/test/openapi.test.ts`: every route registered on the app must appear here and
    every path here must be served by the app. The index also serves this document as JSON at
    `GET /openapi.json`.

    Three kinds of caller use this API:

    - **Phones** (the Identizen app): register with `POST /devices`, then sign every request
      with the `Idz-Signature` header (PROTOCOL.md §8).
    - **Browsers / the SDK**: start logins (`POST /challenge`), run discovery
      (`/discover/*`), and wait on the challenge (`/challenge/{id}/state` or the WebSocket).
    - **Sites** (relying parties): standard OIDC (`/authorize`, `/token`, `/userinfo`), site
      management (`/sites`), and the server-to-server Verification API (`/v1/verify`).

    Errors are JSON `{ "error": "<code>", "error_description": "…" }`; the codes are listed
    per operation and documented at https://docs.identizen.com/errors/. Unknown routes answer
...
    headers are `Content-Type`, `Authorization`, and `Idz-Signature`).
...
servers:
...
paths:
...
  /authorize:
    get:
      tags: [OIDC]
      operationId: authorize
      summary: OIDC authorization endpoint (hosted login page)
      description: |
        Authorization Code flow with PKCE. Creates a challenge session and renders the hosted
        login page (match code, QR, WebSocket to the session). After the phone approves, the
        page redirects the browser to `redirect_uri?code=…&state=…`. The code is single-use and
        bound to the client.

        Validation order: unknown `client_id` or unregistered `redirect_uri` is a JSON `400`
        (nothing is redirected to an unverified URI). Every later failure is an OIDC error
        redirect (`302` to `redirect_uri` with `error`, `error_description`, and `state`):
        `invalid_request` (missing `response_type`, missing or malformed PKCE,
        `acr_values=idz:mfa` alone without `login_hint`, `prompt=none` combined with another value),
        `unsupported_response_type`, `request_not_supported`, `request_uri_not_supported`,
        `invalid_scope` (no `openid`), `interaction_required` (`prompt=none`), and
        `login_required` (no active device is bound to `login_hint`). Unknown scopes and
        unknown `acr_values` are ignored (RFC 6749 §3.3; `acr` is a voluntary claim).

        `acr_values=idz:mfa` with `login_hint=<sub>` is step-up: the challenge is pushed to the
        device bound to `sub`. `login_hint` alone pushes to the bound device at `idz:login`.
        `prompt=enroll` marks an enrollment (the resulting `sub` is what the site stores).
        Rate limited per source IP (`429 rate_limited`) and per client (`429 client_rate_limited`); a step-up whose push would exceed the device's quota is `429 push_rate_limited`.
        The same request may be sent as `POST` with a form body (OpenID Connect Core §3.1.2.1).
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            const: code
        - name: client_id
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ClientId'
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
          description: Must exactly match one of the site's registered `redirect_uris`.
        - name: scope
          in: query
          required: true
          schema:
            type: string
          description: Space-separated; must include `openid`. `handle` releases `idz_handle`.
          example: openid handle
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
        - name: code_challenge_method
          in: query
          required: true
          schema:
            type: string
            const: S256
        - name: state
          in: query
          required: false
          schema:
            type: string
          description: Echoed on the success and error redirects.
        - name: nonce
          in: query
          required: false
          schema:
            type: string
          description: Echoed in the id_token when sent.
        - name: acr_values
          in: query
          required: false
          schema:
            type: string
          description: Space-separated; each value must be `idz:login` or `idz:mfa`. `idz:mfa` requires `login_hint` unless `idz:login` is also listed, in which case the login proceeds at `idz:login`.
        - name: login_hint
          in: query
          required: false
          schema:
            type: string
          description: The per-site `sub` whose bound device should receive the push.
        - name: prompt
          in: query
          required: false
          schema:
            type: string
          description: '`none` always fails with `interaction_required`; `enroll` marks an enrollment. Other values are carried through unchanged.'
      responses:
        '200':
          $ref: '#/components/responses/LoginPage'
        '302':
          $ref: '#/components/responses/AuthorizeErrorRedirect'
        '400':
          $ref: '#/components/responses/AuthorizeBadRequest'
        '429':
          description: '`rate_limited` — per-IP limit; `client_rate_limited` — the site started too many challenges this minute; `push_rate_limited` — a step-up would push to a device that was pushed ten times in the last minute.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [OIDC]
      operationId: authorizePost
      summary: OIDC authorization endpoint (POST form)
      description: |
        Identical to `GET /authorize` with the same parameters in an
        `application/x-www-form-urlencoded` body (OpenID Connect Core §3.1.2.1 requires both).
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AuthorizeRequest'
      responses:
        '200':
          $ref: '#/components/responses/LoginPage'
        '302':
          $ref: '#/components/responses/AuthorizeErrorRedirect'
        '400':
          $ref: '#/components/responses/AuthorizeBadRequest'
        '429':
          description: '`rate_limited` — per-IP limit; `client_rate_limited` — the site started too many challenges this minute; `push_rate_limited` — a step-up would push to a device that was pushed ten times in the last minute.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /challenge:
    post:
      tags: [Challenges]
      operationId: startChallenge
      summary: Start a login (the JSON twin of `/authorize`)
      description: |
        What the SDK calls. Creates a signed challenge in a session and returns what the
        browser needs; the phone fetches the signed challenge itself (`GET /challenge/{id}`).
        When `redirect_uri` is present the session also carries the OIDC parameters so the
        approval yields an authorization code redeemable at `POST /token`; the request is then
        held to exactly the rules of `/authorize` (a registered `redirect_uri`, PKCE S256,
        `scope` including `openid`, `prompt` semantics), with errors returned as JSON
        (`400 invalid_request`, `invalid_scope`, `interaction_required`) instead of a redirect. `acr: idz:mfa` or
        `login_hint` pushes the challenge to the device bound to `login_hint`
        (`400 login_hint_required` when `idz:mfa` has no hint, `400 login_required` when the
        `sub` has no active bound device). Rate limited per source IP and per client.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeStartRequest'
      responses:
        '201':
          description: Challenge created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeStarted'
        '400':
          description: '`invalid_request` — body failed validation, or (with `redirect_uri`) an authorization-request rule failed; `invalid_scope` — `scope` without `openid`; `interaction_required` — `prompt=none`; `login_hint_required`; `login_required` — no active device is bound to `login_hint` for this site.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: '`unknown_client` — no site with that `client_id`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: '`rate_limited` — per-IP limit; `client_rate_limited` — the site started too many challenges this minute; `push_rate_limited` — a step-up would push to a device that was pushed ten times in the last minute.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /challenge/{id}:
    get:
      tags: [Challenges]
      operationId: getChallenge
      summary: The signed challenge (what the phone signs)
      description: Public. The phone verifies `sig` against the pinned index key before showing anything.
      parameters:
        - $ref: '#/components/parameters/ChallengeIdPath'
      responses:
        '200':
          description: Signed challenge plus the session status.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SignedChallenge'
                  - type: object
                    required: [status]
                    properties:
                      status:
                        $ref: '#/components/schemas/ChallengeStatus'
        '404':
          $ref: '#/components/responses/UnknownChallenge'
  /challenge/{id}/state:
    get:
      tags: [Challenges]
      operationId: getChallengeState
      summary: Poll the session (fallback for environments without WebSocket)
      parameters:
        - $ref: '#/components/parameters/ChallengeIdPath'
      responses:
        '200':
          description: Current state. `redirect` is only set once approved.
          content:
            application/json:
              schema:
                type: object
                required: [challenge_id, status, pairing, redirect]
                properties:
                  challenge_id:
                    $ref: '#/components/schemas/ChallengeId'
                  status:
                    $ref: '#/components/schemas/ChallengeStatus'
                  pairing:
                    oneOf:
                      - $ref: '#/components/schemas/SignedPairing'
                      - type: 'null'
                  redirect:
                    type: [string, 'null']
                    format: uri
                    description: '`redirect_uri?code=…&state=…` when the session carried an OIDC request and was approved.'
        '404':
          $ref: '#/components/responses/UnknownChallenge'
  /challenge/{id}/browser-key:
    post:
      tags: [Challenges]
      operationId: attachBrowserKey
      summary: Attach the browser's P-256 public key so approval issues a pairing
      description: 'Used by the hosted login page after render. Accepted only while the session is pending and has no key yet (`ok: false` otherwise).'
      parameters:
        - $ref: '#/components/parameters/ChallengeIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [browser_pubkey]
              properties:
                browser_pubkey:
                  type: string
                  pattern: '^[A-Za-z0-9_-]{80,100}$'
                  description: Raw uncompressed P-256 public key (65 bytes) as base64url.
      responses:
        '200':
          description: Whether the key was attached.
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/UnknownChallenge'
  /challenge/{id}/ws:
    get:
      tags: [Challenges]
      operationId: challengeWebSocket
      summary: WebSocket for the waiting browser
      description: |
        Upgrade to a WebSocket bridged to the session. The server sends JSON text frames:
        `{ "type": "pending", "challenge_id", "exp" }` on connect while pending, then exactly one
        terminal event and a close (code 1000):
        `{ "type": "approved", "challenge_id", "pairing": SignedPairing | null, "redirect": string | null }`,
        `{ "type": "denied", "challenge_id" }`, or `{ "type": "expired", "challenge_id" }`.
        A late joiner receives the terminal event immediately. Sending the text `ping` returns
        `pong`. Without an `Upgrade: websocket` header the response is `426` (plain text).
      parameters:
        - $ref: '#/components/parameters/ChallengeIdPath'
        - name: Upgrade
          in: header
          required: true
          schema:
            type: string
            const: websocket
      responses:
        '101':
          description: Switching Protocols.
        '404':
          description: Unknown challenge (plain text `unknown challenge`).
          content:
            text/plain:
              schema:
                type: string
        '426':
          description: Not a WebSocket upgrade (plain text `expected websocket`).
          content:
            text/plain:
              schema:
                type: string
  /challenge/{id}/assert:
    post:
      tags: [Challenges]
      operationId: submitAssertion
      summary: Submit the phone's double-signed assertion
      description: |
        Device-authenticated (`Idz-Signature`) and double-signed: `device_sig` by the device
        key, `site_sig` by the per-site key. `payload.device_id` must be the signing device.
        Verification order (PROTOCOL.md §4.1): session pending → device active and, when the
        challenge was pushed to a specific device, the same device → assertion schema, challenge
        id, nonce, `rp_id`, `acr`, `reason_hash`, expiry, `iat` window, `device_sig`,
        `site_pubkey`, `sub`, `site_sig` → trust-on-first-use binding of `(rp_id, sub)`.
        Every failure writes a `login.denied` audit event. On success the session is approved,
        the waiting browser is notified, a pairing is issued when the browser supplied a key,
        an OIDC code is minted when the session carried an authorization request, and a
        Verification API record is resolved when the session belongs to one.
      security:
        - idzSignature: []
      parameters:
        - $ref: '#/components/parameters/ChallengeIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignedAssertion'
      responses:
        '200':
          description: Approved.
          content:
            application/json:
              schema:
                type: object
                required: [status, challenge_id, sub, acr, pairing, redirect]
                properties:
                  status:
                    type: string
                    const: approved
                  challenge_id:
                    $ref: '#/components/schemas/ChallengeId'
                  sub:
                    $ref: '#/components/schemas/Sub'
                  acr:
                    $ref: '#/components/schemas/Acr'
                  pairing:
                    oneOf:
                      - $ref: '#/components/schemas/SignedPairing'
                      - type: 'null'
                  redirect:
                    type: [string, 'null']
                    format: uri
        '400':
          description: 'Assertion rejected: `malformed_assertion`, `challenge_mismatch`, `nonce_mismatch`, `rp_id_mismatch`, `acr_mismatch`, `reason_mismatch`, `iat_too_early`, `iat_too_late`, `sub_mismatch`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 'Request signature: `missing_signature`, `unknown_device`, `bad_signature`, `replayed_request`. Assertion: `unknown_device` — `payload.device_id` is not registered; `bad_device_signature`, `bad_site_pubkey`, `bad_site_signature`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: '`device_inactive`; `wrong_device` — `payload.device_id` differs from the signing device, or an untargeted challenge was routed to another device; `wrong_identity` — the challenge was issued for another identity (step-up, Verification API); `wrong_subject` — the assertion `sub` is not the one the challenge was issued for; `insufficient_amr` — an `idz:mfa` assertion whose `amr` names nothing that verified the person.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/UnknownChallenge'
        '409':
          description: '`binding_conflict` — `sub` is already bound to a different site key or identity for this `rp_id`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '410':
          description: '`challenge_approved`, `challenge_denied`, `challenge_expired` — the session already reached that state; `challenge_reserved` — another approval of this challenge is being written; `expired` — the challenge `exp` has passed.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /challenge/{id}/deny:
    post:
      tags: [Challenges]
      operationId: denyChallenge
      summary: Decline a challenge from the phone
      description: Marks a pending session denied and notifies the browser. Only the person the challenge was issued for may decline it (any of their devices); an untargeted challenge that discovery routed to one device is that device's to decline. When the session is no longer pending the current status is returned unchanged. No body.
      security:
        - idzSignature: []
      parameters:
        - $ref: '#/components/parameters/ChallengeIdPath'
      responses:
        '200':
          description: Session status after the call.
          content:
            application/json:
              schema:
                type: object
                required: [status, challenge_id]
                properties:
                  status:
                    $ref: '#/components/schemas/ChallengeStatus'
                  challenge_id:
                    $ref: '#/components/schemas/ChallengeId'
        '401':
          $ref: '#/components/responses/DeviceUnauthorized'
        '403':
          description: '`device_inactive`; `wrong_identity` — the challenge was issued for another identity; `wrong_device` — it was routed to another device.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: '`approval_in_progress` — the assertion for this challenge verified and its writes are in flight; the decline arrived too late to change the outcome.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/UnknownChallenge'
...
  /sites/{client_id}:
    get:
      tags: [Sites]
      operationId: getSite
      summary: Read a site
      description: Bearer token = the site's client secret (HTTP Basic with the secret as password is accepted too). Public clients have no secret and cannot authenticate here.
      security:
        - clientSecretBearer: []
        - clientSecretBasic: []
      parameters:
        - $ref: '#/components/parameters/ClientIdPath'
      responses:
        '200':
          description: The site.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'
        '401':
          $ref: '#/components/responses/InvalidClient'
        '404':
          $ref: '#/components/responses/UnknownClient'
    patch:
      tags: [Sites]
      operationId: updateSite
      summary: Update a site, optionally rotating its secrets
      security:
        - clientSecretBearer: []
        - clientSecretBasic: []
      parameters:
        - $ref: '#/components/parameters/ClientIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SitePatch'
      responses:
        '200':
          description: The updated site. `client_secret` / `webhook_secret` are the new values when rotated, otherwise `null`.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Site'
                  - $ref: '#/components/schemas/SiteSecrets'
        '400':
          description: '`invalid_request` — body failed validation; `invalid_destination` — `webhook_url` or `backchannel_logout_uri` is one the index will not call (not https, credentials in the URL, or a local, private or link-local address).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/InvalidClient'
        '404':
          $ref: '#/components/responses/UnknownClient'
      callbacks:
        backchannelLogout:
          $ref: '#/components/callbacks/BackchannelLogout'
        verificationWebhook:
          $ref: '#/components/callbacks/VerificationWebhook'
...
components:
  securitySchemes:
    accessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The OIDC access token from `POST /token` (`typ: at+jwt`, ES256, 1 hour, audience =
        the index). `/userinfo` accepts any client's token. `/me/*` additionally requires the
        token to have been issued to a client listed in the index's `DASHBOARD_CLIENT_IDS`
        (`403 not_dashboard_client`) and checks the session for revocation on every call.
    idzSignature:
      type: apiKey
      in: header
      name: Idz-Signature
      description: |
        Device request authentication (PROTOCOL.md §8):

            Idz-Signature: v1,d=<device_id>,t=<unix seconds>,s=<base64url sig>

        where `sig = Ed25519.sign(deviceKey, UTF8("identizen/v1/request\n" + METHOD + "\n" +
        PATH + "\n" + base64url(SHA-256(body)) + "\n" + t))`. `METHOD` is upper-case, `PATH`
        is the request path including the query string, `body` is the raw request body (empty
        string for none). The index rejects `t` outside ±60 s of its clock (`401 bad_signature`)
        and any `(device_id, t, sig)` seen before (`401 replayed_request`); a device may send at
        most 120 signed requests per minute. `@identizen/protocol` exports `signRequest`.
    clientSecretBearer:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <client_secret>`. On `/sites/{client_id}` the client id comes
        from the path; on `/v1/verify*` it must be sent in the `Idz-Client-Id` header.
    clientSecretBasic:
      type: http
      scheme: basic
      description: '`Authorization: Basic base64(client_id:client_secret)`.'
    siteRegistrationToken:
      type: http
      scheme: bearer
      description: '`Authorization: Bearer <SITE_REGISTRATION_TOKEN>`; required by `POST /sites` unless the index runs with `OPEN_SITE_REGISTRATION=true`.'
  parameters:
...
    ChallengeIdPath:
      name: id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ChallengeId'
...
  responses:
...
    LoginPage:
      description: The hosted login page.
      headers:
        Cache-Control:
          schema:
            type: string
            const: no-store
        X-Frame-Options:
          schema:
            type: string
            const: DENY
        Referrer-Policy:
          schema:
            type: string
            const: no-referrer
      content:
        text/html:
          schema:
            type: string
    AuthorizeErrorRedirect:
      description: OIDC error redirect to `redirect_uri` with `error`, `error_description`, and `state` query parameters.
      headers:
        Location:
          schema:
            type: string
            format: uri
    AuthorizeBadRequest:
      description: '`invalid_client` — unknown `client_id`; `invalid_request` — `redirect_uri` is not registered for this client.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      description: Every error body. `issues` is present only for `invalid_request` from body validation.
      required: [error, error_description]
      properties:
        error:
          type: string
          description: Stable machine-readable code; see https://docs.identizen.com/errors/.
        error_description:
          type: string
        issues:
          type: array
          items:
            type: object
            required: [path, message]
            properties:
              path:
                type: string
                description: Dotted path of the offending field.
              message:
                type: string
...
    ClientId:
      type: string
      description: '`idz_live_<ULID>` or `idz_test_<ULID>`.'
      pattern: '^idz_(live|test)_[0-9A-HJKMNP-TV-Z]{26}$'
...
    ChallengeStatus:
      type: string
      description: Session status of a challenge.
      enum: [pending, approved, denied, expired]
...
    SignedChallenge:
      type: object
      additionalProperties: false
      required: [payload, sig]
      properties:
        payload:
          $ref: '#/components/schemas/Challenge'
        sig:
          $ref: '#/components/schemas/Signature'
          description: Index Ed25519 signature of type `challenge` over the canonical payload.
...